Upload training script (WIP pynative)

+++ src/models/utils/geometry.py
@@ -1,3 +1,4 @@
+import mindspore as ms
 from mindspore import ops
 
 
@@ -39,8 +40,8 @@ def warp_kpts(kpts0, depth0, depth1, T_0to1, K0, K1):
 
     # Covisible Check
     h, w = depth1.shape[1:3]
-    covisible_mask = (w_kpts0[:, :, 0] > 0) * (w_kpts0[:, :, 0] < w-1) * \
-        (w_kpts0[:, :, 1] > 0) * (w_kpts0[:, :, 1] < h-1)
+    covisible_mask = (w_kpts0[:, :, 0] > 0).astype(ms.int32) * (w_kpts0[:, :, 0] < w-1).astype(ms.int32) * \
+        (w_kpts0[:, :, 1] > 0).astype(ms.int32) * (w_kpts0[:, :, 1] < h-1).astype(ms.int32)
     w_kpts0_long = w_kpts0.long()
     w_kpts0_long[~covisible_mask, :] = 0
 
@@ -48,6 +49,6 @@ def warp_kpts(kpts0, depth0, depth1, T_0to1, K0, K1):
         [depth1[i, w_kpts0_long[i, :, 1], w_kpts0_long[i, :, 0]] for i in range(w_kpts0_long.shape[0])], axis=0
     )  # (N, L)
     consistent_mask = ((w_kpts0_depth - w_kpts0_depth_computed) / w_kpts0_depth).abs() < 0.2
-    valid_mask = nonzero_mask * covisible_mask * consistent_mask
+    valid_mask = nonzero_mask.astype(ms.int32) * covisible_mask * consistent_mask.astype(ms.int32)
 
     return valid_mask, w_kpts0
